0619 수업끝 커밋 개인 작업 완료. xml 디자인 해야함.
