+++ app/src/main/AndroidManifest.xml
@@ -18,16 +18,19 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
+            android:name=".WaitinglistActivity"
+            android:exported="false" />
+        <activity
             android:name=".waitinglistActivity"
             android:exported="false" />
         <activity
-            android:name=".CategoryOneActivity"
+            android:name=".category.CategoryOneActivity"
             android:exported="true" />
         <activity
-            android:name=".CategoryThreeActivity"
+            android:name=".category.CategoryThreeActivity"
             android:exported="true" />
         <activity
-            android:name=".CategoryTwoActivity"
+            android:name=".category.CategoryTwoActivity"
             android:exported="true" />
         <activity
             android:name=".DetailActivity"
