--- app/src/main/java/com/example/a1st_web_app_project/fragment/BlankFragment.kt
@@ -1,14 +1,0 @@
-package com.example.a1st_web_app_project.fragment
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.example.a1st_web_app_project.R
-
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
-
