+++ app/src/main/res/layout/activity_categorytwo.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CategoryTwoActivity">
+    tools:context=".category.CategoryTwoActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar"
@@ -29,7 +29,7 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.0"
-        tools:context=".CategoryTwoActivity">
+        tools:context=".category.CategoryTwoActivity">
 
 
         <LinearLayout
@@ -38,7 +38,7 @@
             android:layout_marginTop="60dp"
             android:gravity="center_horizontal"
             android:orientation="horizontal"
-            tools:context=".CategoryTwoActivity">
+            tools:context=".category.CategoryTwoActivity">
 
 
             <androidx.appcompat.widget.AppCompatButton
