--- app/src/main/res/layout/categoryfrag_1p.xml
@@ -1,104 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".CategoryTwoActivity">
-
-    <androidx.appcompat.widget.Toolbar
-        android:id="@+id/toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="#FFAC07"
-        android:theme="@style/ThemeOverlay.AppCompat.ActionBar"
-        app:layout_constraintTop_toTopOf="parent"
-        app:popupTheme="@style/ThemeOverlay.AppCompat.Light"
-        tools:layout_editor_absoluteX="0dp">
-
-
-    </androidx.appcompat.widget.Toolbar>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.0"
-        tools:context=".WaitingActivity">
-
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="89dp"
-            android:layout_marginTop="60dp"
-            android:gravity="center_horizontal"
-            android:orientation="horizontal"
-            tools:context=".WaitingActivity">
-
-
-            <androidx.appcompat.widget.AppCompatButton
-                android:id="@+id/catebtn1"
-                android:layout_width="100dp"
-                android:layout_height="60dp"
-                android:layout_marginTop="18dp"
-                android:layout_marginRight="20dp"
-                android:background="@drawable/white_radius"
-                android:text="중구"
-                android:textSize="20dp"
-                android:textStyle="bold" />
-
-            <androidx.appcompat.widget.AppCompatButton
-                android:id="@+id/catebtn2"
-                android:layout_width="100dp"
-                android:layout_height="60dp"
-                android:layout_marginTop="18dp"
-                android:background="@drawable/white_radius"
-                android:text="종로구"
-                android:textSize="18dp"
-                android:textStyle="bold" />
-
-            <androidx.appcompat.widget.AppCompatButton
-                android:id="@+id/catebtn3"
-                android:layout_width="100dp"
-                android:layout_height="60dp"
-                android:layout_marginLeft="20dp"
-                android:layout_marginTop="18dp"
-                android:background="@drawable/white_radius"
-                android:text="기타"
-                android:textSize="20dp"
-                android:textStyle="bold" />
-        </LinearLayout>
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/recyclerView"
-            android:layout_width="360dp"
-            android:layout_height="match_parent"
-            android:orientation="vertical"
-            android:layout_gravity="center_horizontal"
-            app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintVertical_bias="0.496"
-            tools:ignore="MissingConstraints"
-            tools:layout_editor_absoluteX="16dp" />
-    </LinearLayout>
-
-    <com.google.android.material.bottomnavigation.BottomNavigationView
-        android:id="@+id/bottommenu"
-        android:layout_width="match_parent"
-        android:layout_height="50dp"
-        android:layout_gravity="bottom"
-        android:background="#FF9800"
-        app:itemIconTint="#000000"
-        app:labelVisibilityMode="unlabeled"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:menu="@menu/menu"
-        tools:layout_editor_absoluteX="-16dp" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
-
