--- app/src/main/res/layout/categoryfrag_2p.xml
@@ -1,80 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".MainActivity2">
-
-
-    <androidx.appcompat.widget.Toolbar
-        android:id="@+id/toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="?attr/actionBarSize"
-        android:background="#FFAC07"
-        android:theme="@style/ThemeOverlay.AppCompat.ActionBar"
-        app:layout_constraintTop_toTopOf="parent"
-        app:popupTheme="@style/ThemeOverlay.AppCompat.Light"
-        tools:layout_editor_absoluteX="0dp" />
-
-    <LinearLayout
-        android:layout_width="413dp"
-        android:layout_height="94dp"
-        android:gravity="center_horizontal"
-        android:orientation="horizontal"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        tools:layout_editor_absoluteY="62dp"
-        tools:ignore="MissingConstraints">
-
-        <Button
-            android:id="@+id/sbtn1"
-            android:layout_width="100dp"
-            android:layout_height="60dp"
-            android:layout_marginTop="18dp"
-            android:layout_marginRight="20dp"
-            android:background="@drawable/white_radius"
-            android:drawableTop="@drawable/best"
-            android:text="인기순위"
-            android:textSize="20dp"
-            android:textStyle="bold" />
-
-        <Button
-            android:id="@+id/sbtn2"
-            android:layout_width="100dp"
-            android:layout_height="60dp"
-            android:layout_marginTop="18dp"
-            android:background="@drawable/white_radius"
-            android:drawableTop="@drawable/anniversary"
-            android:text="동네별"
-            android:textSize="20dp"
-            android:textStyle="bold" />
-
-        <Button
-            android:id="@+id/sbtn3"
-            android:layout_width="100dp"
-            android:layout_height="60dp"
-            android:layout_marginTop="18dp"
-            android:layout_marginLeft="20dp"
-            android:background="@drawable/white_radius"
-            android:drawableTop="@drawable/dining"
-            android:text="장르별"
-            android:textSize="20dp"
-            android:textStyle="bold" />
-
-    </LinearLayout>
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recyclerView"
-        android:layout_width="match_parent"
-        android:layout_height="500dp"
-        android:orientation="vertical"
-        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.987"
-        tools:ignore="MissingConstraints" />
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
