--- app/src/main/res/layout/fragment_container.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".FilterFragmentActivity">
-
-    <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical">
-
-        <Button
-            android:id="@+id/button_filter1"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Filter 1" />
-
-        <Button
-            android:id="@+id/button_filter2"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Filter 2" />
-
-    </LinearLayout>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
