+++ app/src/main/res/layout/item_main.xml
@@ -54,11 +54,31 @@
                 android:id="@+id/contactView3"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content" />
+            <LinearLayout
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="10dp"
+                android:gravity="center_vertical">
 
-            <TextView
-                android:id="@+id/contactView4"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content" />
+                <TextView
+                    android:id="@+id/contactView4"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:textAlignment="center"
+                    />
+
+                <ImageView
+                    android:id="@+id/imageViewPopularityIcon"
+                    android:layout_width="32dp"
+                    android:layout_height="34dp"
+                    android:layout_marginBottom="5dp"
+                    android:layout_marginStart="5dp"
+                    android:src="@drawable/beststar" />
+            </LinearLayout>
+<!--            <TextView-->
+<!--                android:id="@+id/contactView4"-->
+<!--                android:layout_width="wrap_content"-->
+<!--                android:layout_height="wrap_content" />-->
 
         </LinearLayout>
     </LinearLayout>
